RDK-35046 : Slow Motion Playback in AAMP
Reason for change: Added code changes to support slow motion playback and added supported slow motion playback speed in Mediametadata event.
Test Procedure: Any hls/dash mp4 fragmented file should be able to play in slow motion.
Risks: Low

DEPS_BRANCH=rdk/components/generic/gst-plugins-rdk-aamp:dev_sprint_22_1
DEPS_BRANCH=rdk/components/generic/aampabr:dev_sprint_22_1
DEPS_BRANCH=rdk/components/generic/aampmetrics:dev_sprint_22_1

+++ AampUtils.cpp
@@ -961,10 +961,10 @@ bool aamp_WriteFile(std::string fileName, const char* data, size_t len, MediaTyp
 /**
  * @brief Get compatible trickplay for 6s cadense of iframe track from the given rates
  */
-int getWorkingTrickplayRate(int rate)
+float getWorkingTrickplayRate(float rate)
 {
-	int workingRate;
-	switch (rate){
+	float workingRate;
+	switch ((int)rate){
 		case 4:
 			workingRate = 25;
 			break;
@@ -992,10 +992,10 @@ int getWorkingTrickplayRate(int rate)
 /**
  * @brief Get reverse map the working rates to the rates given by platform player
  */
-int getPseudoTrickplayRate(int rate)
+float getPseudoTrickplayRate(float rate)
 {
-	int psudoRate;
-	switch (rate){
+	float psudoRate;
+	switch ((int)rate){
 		case 25:
 			psudoRate = 4;
 			break;
